Fixing the grid system
Modern browsers support the "border-box" value for the "box-sizing" CSS
property, which allows us to accurately compute elements' widths, as
long as we leave margins out of the way. I thus modified the grid system
to use paddings on the horizontal direction, and in conjunction with
"border-box", it let me switch to meaningful widths (no more 29,3333…%).

I leveraged this enhanced grid system to "box out" the two implicit columns
you have in the hero section, using .one-third and a new .two-third class.

@see https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing
@see https://css-tricks.com/almanac/properties/b/box-sizing/

Signed-off-by: Jean-Denis Vauguet <jd@vauguet.fr>